Restructuring and finalizing parts
