+++ SpiceSharp/Simulations/IntegrationMethods/Breakpoints/Breakpoints.cs
