+++ SpiceSharp/Simulations/IntegrationMethods/Spice/StateHistoryDefault.cs
